requirements.txt
Flask==2.1.2
numpy==1.21.6
Pillow==9.1.1
streamlit==1.10.0
tensorflow==2.9.1
